+++ Changelog.md
@@ -1,4 +1,5 @@
 ## 0.0.59
++ Added `muna.Sandbox.run_commands` function to run shell commands when setting up the compiler sandbox.
 + Fixed CLI prediction inputs with hyphen characters not being parsed correctly.
 
 ## 0.0.58